+++ src/App.vue
@@ -27,6 +27,7 @@
         flex-wrap: wrap;
     }
 
+
     h2 {
         width:100%;
         text-align:center;
@@ -37,6 +38,33 @@
         font-size: 200%;
     }
 
-    * {
+    .table {
+        width:100%;
     }
+
+    .table-striped th {
+        border-bottom: 2px solid rgba(55, 27, 9, 0.83);
+    }
+
+    .table td {
+        padding: 5px;
+    }
+
+    .table-striped {
+        border-collapse: collapse;
+    }
+
+    .table-striped tr:nth-child(even) td{
+        border-top: 2px solid rgba(175, 121, 8, 0.52);
+        border-bottom: 2px solid rgba(175, 121, 8, 0.52);
+    }
+
+    .table-striped tr:nth-child(even) {
+        background-color: rgba(190, 165, 32, 0.73);
+    }
+
+    .table-striped tr:nth-child(odd) {
+        background-color: rgba(255, 255, 255, 0.75);
+    }
+
 </style>